Reimplement FT removal. (#101)
