+++ fasttrips/Assignment.py
@@ -1741,8 +1741,9 @@ class Assignment:
         ######################################################################################################
         FastTripsLogger.info("  Step 2. Calculate costs and probabilities for all pathset paths")
         (pathset_paths_df, pathset_links_df) = PathSet.calculate_cost(
-            FT, Assignment.STOCH_DISPERSION, pathset_paths_df, pathset_links_df,
-            veh_trips_df, reset_bump_iter=simulation_iteration==0)
+            Assignment.STOCH_DISPERSION, pathset_paths_df, pathset_links_df, FT.veh_trips_df,
+            FT.passengers.trip_list_df, FT.routes, FT.tazs, FT.transfers, stops=FT.stops,
+            reset_bump_iter=simulation_iteration==0)
 
         ######################################################################################################
         FastTripsLogger.info("  Step 3. Choose a path for each passenger from their pathset")
@@ -1807,9 +1808,16 @@ class Assignment:
 
             ######################################################################################################
             FastTripsLogger.info("  Step 3. Calculate costs and probabilities for all pathset paths")
+            pathset_paths_df.to_csv('C:/apps/fast-trips/fasttrips/Examples/misc/calculate_cost/paths_i{}_pf{}_s{}.csv'.format(iteration, pathfinding_iteration, simulation_iteration),
+                                    index=False)
+            pathset_links_df.to_csv('C:/apps/fast-trips/fasttrips/Examples/misc/calculate_cost/links_i{}_pf{}_s{}.csv'.format(iteration, pathfinding_iteration, simulation_iteration),
+                                    index=False)
+            veh_trips_df.to_csv('C:/apps/fast-trips/fasttrips/Examples/misc/calculate_cost/vehicles_i{}_pf{}_s{}.csv'.format(iteration, pathfinding_iteration, simulation_iteration),
+                                    index=False)
             (pathset_paths_df, pathset_links_df) = PathSet.calculate_cost(
-                FT, Assignment.STOCH_DISPERSION, pathset_paths_df, pathset_links_df,
-                veh_trips_df, reset_bump_iter=simulation_iteration == 0)
+                Assignment.STOCH_DISPERSION, pathset_paths_df, pathset_links_df, veh_trips_df,
+                FT.passengers.trip_list_df, FT.routes, FT.tazs, FT.transfers, stops=FT.stops,
+                reset_bump_iter=simulation_iteration == 0)
 
             ######################################################################################################
             FastTripsLogger.info("  Step 4. Choose a path for each passenger from their pathset")
